+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 **Note:** Patch versions prior to [v8.0.0](#800---2018-07-01) are not listed here.
 
+## [11.0.1] - 2018-10-21
+### Changed
+- Moved jsdoc and minami to dev dependencies
+- Use `chalk` instead of `colors`
+
 ## [11.0.0] - 2018-10-21
 ### Added
 - :tada: **Documentation!** :tada: